backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-dotenv
pydantic[email]
bcrypt
python-jose[cryptography]
pyotp
fastapi-mail
redis
cloudinary
pdfplumber
python-multipart
celery
flower
groq
python-docx